RubberDuck debugging iOS client.
This is the iOS client of our Alexa based smart RubberDuck.
This was our entry into the Swamphacks 2018 hackathon.

The iOS client records users voice input via the microphone and
upoads the audio file to AVS - Amazon voice service. We had created
custom a skill for our alexa which would search stackoverflow for
bugs matching the description in the audio recorded by the user.

The results get sent back down to the iOS app as a audio file which
we can playback to the user.